add spark modules to hermione_spark module template
