+++ hermione/cli.py
@@ -53,7 +53,7 @@ def new(project_name, implemented):
         }
     os.makedirs(os.path.join(LOCAL_PATH, project_name))
     if is_imp:
-        option = click.prompt('Do you want to start with: \n\t(1) Sagemaker \n\t(2) PySpark \n\t(3) Local version \n', type=int, default=3)
+        option = click.prompt('Do you want to start with: \n\t(1) Sagemaker \n\t(2) Local version \n', type=int, default=3)
         implemented_version_type(project_name,custom_inputs,option)
     else:
         write_module(os.path.join(LOCAL_PATH, project_name), '__NOT_IMPLEMENTED_BASE__', True, custom_inputs)
@@ -73,8 +73,6 @@ def implemented_version_type(project_name,custom_inputs,option):
     """
     if option == 1:
         write_module(os.path.join(LOCAL_PATH, project_name), '__IMPLEMENTED_SAGEMAKER__', True, custom_inputs)
-    if option == 2:
-        write_module(os.path.join(LOCAL_PATH, project_name), '__IMPLEMENTED_PYSPARK__', True, custom_inputs)
     else:
         write_module(os.path.join(LOCAL_PATH, project_name), '__IMPLEMENTED_BASE__', True, custom_inputs)
