--- hermione/module_templates/__IMPLEMENTED_PYSPARK__/.tpl.gitignore
@@ -1,8 +1,0 @@
-.ipynb_checkpoints 
-mlruns/ 
-__pycache__/ 
-.vscode/ 
-catboost_info/ 
-.metaflow 
-data/ 
-*_env/
