--- hermione/module_templates/__IMPLEMENTED_PYSPARK__/data/raw/raw_test.csv
@@ -1,269 +1,0 @@
-Survived,Pclass,Sex,Age
-1,3,female,1.0
-0,3,male,
-0,1,male,30.0
-0,1,male,61.0
-0,3,male,27.0
-0,2,male,46.0
-1,2,female,40.0
-0,2,female,27.0
-0,3,female,18.0
-0,3,female,29.0
-0,3,female,9.0
-0,3,female,28.0
-1,1,female,52.0
-0,3,male,24.5
-0,1,female,50.0
-1,3,female,0.75
-1,1,female,58.0
-0,1,male,
-0,3,female,45.0
-0,2,male,29.0
-0,3,female,
-1,2,female,8.0
-0,2,male,39.0
-0,3,male,4.0
-0,3,male,2.0
-0,3,male,25.0
-0,3,male,22.0
-1,1,female,17.0
-0,3,male,19.0
-0,3,male,
-0,3,female,2.0
-1,2,female,40.0
-0,2,male,34.0
-1,3,male,26.0
-0,2,male,19.0
-0,3,male,11.0
-0,3,male,42.0
-0,3,male,51.0
-1,2,female,24.0
-1,1,male,40.0
-1,1,female,14.0
-0,3,male,
-0,3,female,
-1,3,female,63.0
-0,3,male,16.0
-0,2,male,25.0
-1,1,female,39.0
-0,1,male,42.0
-0,3,male,20.0
-1,2,female,24.0
-1,3,female,
-0,3,female,6.0
-0,3,male,20.5
-0,3,male,35.0
-0,2,male,24.0
-0,3,male,
-0,3,male,16.0
-0,3,male,18.0
-0,1,male,29.0
-0,3,male,14.0
-1,1,female,33.0
-1,1,female,18.0
-1,1,male,11.0
-1,1,female,
-0,3,male,24.0
-0,3,male,34.0
-0,2,male,48.0
-0,2,male,50.0
-0,1,male,
-1,2,female,
-1,1,female,49.0
-0,3,male,50.0
-1,1,male,
-0,1,male,65.0
-0,3,male,21.0
-0,3,male,28.0
-0,3,male,41.0
-0,3,male,21.0
-1,1,female,
-1,1,female,
-0,3,female,26.0
-0,3,male,28.5
-0,3,male,9.0
-1,3,male,
-0,3,male,24.0
-0,2,male,33.0
-1,3,female,1.0
-0,3,male,33.0
-1,1,male,
-0,3,male,25.0
-1,3,female,18.0
-1,2,male,
-0,2,male,54.0
-1,3,male,3.0
-0,1,male,37.0
-0,3,male,19.0
-1,1,female,24.0
-1,1,female,35.0
-0,3,male,33.0
-1,2,female,24.0
-0,3,male,1.0
-0,1,male,58.0
-0,1,male,45.0
-1,3,female,15.0
-0,3,male,
-0,1,male,31.0
-0,3,male,26.0
-0,3,male,28.5
-0,3,male,35.0
-1,2,female,36.0
-1,2,male,0.83
-1,1,male,31.0
-1,1,female,31.0
-0,3,male,32.0
-0,3,male,26.0
-0,2,male,44.0
-0,1,male,60.0
-0,2,male,54.0
-0,3,male,18.0
-0,1,male,19.0
-0,3,male,19.0
-0,3,male,43.0
-1,1,male,42.0
-0,3,male,
-0,2,male,27.0
-0,2,male,21.0
-0,3,female,43.0
-0,1,male,
-1,3,female,
-0,3,male,20.0
-1,1,female,44.0
-0,3,male,29.0
-0,3,male,25.0
-1,2,female,30.0
-1,1,female,48.0
-0,3,male,18.0
-1,1,female,50.0
-0,3,female,30.0
-0,3,male,
-1,2,female,41.0
-0,3,male,65.0
-1,1,male,32.0
-0,3,male,17.0
-1,2,male,0.83
-1,3,female,5.0
-0,3,male,28.0
-0,3,male,39.0
-1,3,male,32.0
-1,3,male,27.0
-1,1,female,24.0
-1,1,male,49.0
-0,3,male,11.0
-0,3,female,23.0
-0,3,male,16.0
-1,1,female,22.0
-0,3,male,
-1,1,female,58.0
-1,3,female,5.0
-1,2,female,42.0
-1,1,female,49.0
-1,2,female,40.0
-0,1,male,38.0
-1,1,male,25.0
-0,1,female,25.0
-0,1,male,
-1,1,female,22.0
-1,2,female,54.0
-0,3,male,20.0
-1,3,female,
-0,3,male,40.0
-0,3,male,23.0
-0,1,male,
-0,3,female,
-0,2,male,28.0
-0,3,male,
-1,3,male,45.0
-1,2,male,1.0
-0,3,male,
-0,2,male,27.0
-1,1,female,16.0
-0,3,male,31.0
-0,3,male,45.5
-0,3,male,21.0
-1,1,male,23.0
-0,2,male,52.0
-1,2,female,3.0
-0,2,male,16.0
-1,1,female,30.0
-0,3,male,21.0
-1,3,female,16.0
-1,2,male,19.0
-0,3,male,
-1,2,female,25.0
-1,2,male,32.0
-0,3,female,30.5
-0,1,male,21.0
-1,3,male,25.0
-1,3,female,35.0
-1,2,female,17.0
-1,3,male,
-1,1,female,16.0
-1,3,female,
-1,1,male,42.0
-0,1,male,62.0
-0,1,male,40.0
-0,3,male,19.0
-0,3,male,
-1,2,female,25.0
-0,1,male,
-0,3,male,42.0
-0,3,male,29.0
-1,2,female,19.0
-0,3,female,3.0
-0,3,male,35.0
-0,3,male,30.5
-0,3,male,
-0,2,male,34.0
-0,3,male,
-1,3,male,
-0,3,male,25.0
-0,3,male,
-1,1,male,60.0
-0,3,male,30.0
-1,1,male,50.0
-0,3,female,9.0
-0,3,male,25.0
-0,2,male,27.0
-0,3,male,40.5
-0,2,male,30.0
-1,2,female,28.0
-1,3,female,30.0
-0,3,male,74.0
-0,3,female,25.0
-0,3,male,34.0
-0,2,male,31.0
-1,3,male,27.0
-1,1,female,35.0
-0,3,male,7.0
-1,1,female,
-0,2,male,23.0
-0,3,male,30.0
-0,3,male,2.0
-1,3,female,24.0
-0,3,male,
-0,3,male,
-0,2,male,59.0
-0,2,male,51.0
-0,1,male,22.0
-1,1,male,34.0
-1,3,female,33.0
-0,3,male,24.0
-0,3,female,47.0
-0,1,male,47.0
-0,2,male,36.0
-0,3,male,
-1,2,female,14.0
-0,3,female,41.0
-0,1,male,
-1,3,female,
-0,1,male,
-0,3,male,33.0
-1,3,female,31.0
-0,3,male,17.0
-0,3,male,19.0
-0,3,female,2.0
-1,2,female,18.0
-0,3,male,
-1,1,female,52.0
