--- hermione/module_templates/__IMPLEMENTED_PYSPARK__/requirements.txt
@@ -1,29 +1,0 @@
-category-encoders
-coverage
-datetime
-Flask
-gunicorn
-hermione-ml
-matplotlib
-mlflow
-mlxtend
-numpy
-pandas
-plotly
-pytest
-seaborn
-scikit-learn
-scipy
-statsmodels
-tqdm
-yellowbrick
-vega_datasets
-altair
-pandas_profiling
-streamlit_pandas_profiling
-interpret-community
-lime
-lightgbm
-great_expectations
-ipykernel
-pyspark
