+++ hermione/module_templates/hermione_spark.json
@@ -1,0 +1,8 @@
+{
+    "info": "Module for developing Spark classes",
+    "input_info": [],
+    "developers": [
+        "pedro.toledo@a3data.com.br"
+    ],
+    "version": "0.0.1"
+}
