+++ hermione/module_templates/hermione_spark/hermione_spark/__init__.py
@@ -1,0 +1,9 @@
+from . import preprocessing, analysis, model, data_source, utils
+
+__all__ = [
+    "preprocessing", 
+    'analysis', 
+    'data_source', 
+    'model',
+    'utils'
+]
