+++ hermione/module_templates/hermione_spark/hermione_spark/analysis/__init__.py
@@ -1,0 +1,10 @@
+from ._cluster import SparkCluster
+from ._pca import SparkPCA
+from ._feature_selection import SparkFS
+
+
+__all__ = [
+    'SparkCluster',
+    'SparkPCA',
+    'SparkFS'
+]
