+++ hermione/module_templates/hermione_spark/hermione_spark/data_source/__init__.py
@@ -1,0 +1,10 @@
+from ._cleaning import SparkCleaner
+from ._database import SparkDataBase
+from ._spreadsheet import SparkSpreadsheet
+
+
+__all__ = [
+    'SparkSpreadsheet',
+    'SparkDataBase',
+    'SparkCleaner'
+]
