+++ hermione/module_templates/hermione_spark/hermione_spark/main.py
@@ -1,0 +1,13 @@
+from pyspark.sql import SparkSession
+from .data_source import SparkCleaner
+
+if __name__ == "__main__":
+    spark = (
+        SparkSession
+        .builder
+        .config("spark.serializer", "org.apache.spark.serializer.KryoSerializer")
+        .getOrCreate()
+    )
+    cleaner = SparkCleaner(spark)
+    cleaner.clean()
+    spark.stop()
