+++ hermione/module_templates/hermione_spark/hermione_spark/model/__init__.py
@@ -1,0 +1,13 @@
+from ._metrics import SparkMetrics, CustomBinaryEvaluator, CustomRegressionEvaluator
+from ._trainer import SparkTrainer, SparkUnsupTrainer
+from ._wrapper import SparkWrapper
+
+
+__all__ = [
+    'SparkMetrics', 
+    'CustomBinaryEvaluator', 
+    'CustomRegressionEvaluator',  
+    'SparkTrainer', 
+    'SparkUnsupTrainer', 
+    'SparkWrapper'
+]
