+++ hermione/module_templates/hermione_spark/hermione_spark/preprocessing/__init__.py
@@ -1,0 +1,10 @@
+from ._normalization import SparkScaler
+from ._preprocessing import SparkPreprocessor
+from ._text_vectorizer import SparkVectorizer
+
+
+__all__ = [
+    'SparkScaler',
+    'SparkPreprocessor',
+    'SparkVectorizer'
+]
